Move to common namespace
This commit does several things at once:

- Move the server python bits to the common namespace.
- Move s3backup to the common namespace.
- Move the configtools to the common namespace.
- Update unittests
- Fix typo in requirements.txt
- Remove $PYTHONPATH from shell scripts.
