+++ server/bin/getconf.py
@@ -2,7 +2,7 @@
 
 import sys
 
-import configtools
+from pbench.lib import configtools
 
 if __name__ == "__main__":
     opts, args = configtools.parse_args(
